Swap out two buttons for dual-use button
Instead of separate "Brew EE" and "Use RR" buttons, this has a single
button per cauldron that switches between the two based on if the
cauldron is full or not. Buttons now use images instead of words and
show spinners while performing the operation. The buttons are also
positioned better and do not mess up the UI. Bonus, the queued items
are better aligned with the image.